Add task 4.2, add checks for existing for task 4.1

+++ 4/1.sql
@@ -1,13 +1,16 @@
 USE AdventureWorks2012;
 GO
 
-DROP TABLE [Production].[LocationHost];
+IF OBJECT_ID('[Production].[LocationHost]') IS NOT NULL
+  DROP TABLE [Production].[LocationHost];
 GO
 
-DROP TRIGGER [Production].[Trigger_Location_After];
+IF OBJECT_ID('[Production].[Trigger_Location_After]') IS NOT NULL
+  DROP TRIGGER [Production].[Trigger_Location_After];
 GO
 
-DROP VIEW [Production].[View_Location];
+IF OBJECT_ID('[Production].[View_Location]') IS NOT NULL
+  DROP VIEW [Production].[View_Location];
 GO
 
 DELETE FROM [Production].[Location]
@@ -15,60 +18,65 @@ WHERE [Name] = 'Test Location';
 
 -- a
 CREATE TABLE [Production].[LocationHost] (
-	ID INT IDENTITY(1, 1) PRIMARY KEY,
-	Action CHAR(6) NOT NULL CHECK (Action IN('insert', 'update', 'delete')),
-    ModifiedDate DATETIME NOT NULL,
-    SourceID INT NOT NULL,
-    UserName VARCHAR(50) NOT NULL
+  ID           INT IDENTITY (1, 1) PRIMARY KEY,
+  Action       CHAR(6)     NOT NULL CHECK (Action IN ('insert', 'update', 'delete')),
+  ModifiedDate DATETIME    NOT NULL,
+  SourceID     INT         NOT NULL,
+  UserName     VARCHAR(50) NOT NULL
 );
 GO
 
 
 -- b
 CREATE TRIGGER [Production].[Trigger_Location_After]
-ON [Production].[Location]
-AFTER INSERT, UPDATE, DELETE AS 
-    INSERT INTO [Production].[LocationHost](Action, ModifiedDate, SourceID, UserName) 
+  ON [Production].[Location]
+AFTER INSERT, UPDATE, DELETE AS
+  INSERT INTO [Production].[LocationHost] (Action, ModifiedDate, SourceID, UserName)
     SELECT
-      CASE WHEN inserted.LocationID IS NULL THEN 'delete'
-           WHEN  deleted.LocationID IS NULL THEN 'insert'
-		   ELSE 'update'
-      END                                                   AS Action,
-      GETDATE()                                             AS ModifiedDate,
-	  COALESCE(inserted.LocationID, deleted.LocationID) AS SourceID,
-      USER_NAME()                                           AS UserName
-    FROM inserted FULL OUTER JOIN deleted
-    ON inserted.LocationID = deleted.LocationID
+      CASE WHEN inserted.LocationID IS NULL
+        THEN 'delete'
+      WHEN deleted.LocationID IS NULL
+        THEN 'insert'
+      ELSE 'update'
+      END                                               AS Action,
+      GETDATE()                                         AS ModifiedDate,
+      COALESCE(inserted.LocationID, deleted.LocationID) AS SourceID,
+      USER_NAME()                                       AS UserName
+    FROM inserted
+      FULL OUTER JOIN deleted
+        ON inserted.LocationID = deleted.LocationID
 GO
 
 
 -- c
-CREATE VIEW [Production].[View_Location] AS SELECT * FROM [Production].[Location]
+CREATE VIEW [Production].[View_Location] AS
+  SELECT *
+  FROM [Production].[Location]
 GO
 
 
 --d
 INSERT INTO [Production].[View_Location]
-           ([Name]
-           ,[CostRate]
-           ,[Availability]
-           ,[ModifiedDate])
-     VALUES
-           ('Test Location'
-           ,'123' 
-           ,'1.1' 
-           ,GETDATE())
+([Name]
+  , [CostRate]
+  , [Availability]
+  , [ModifiedDate])
+VALUES
+  ('Test Location'
+    , '123'
+    , '1.1'
+    , GETDATE())
 GO
 
 
-
 UPDATE [Production].[Location]
-   SET [Name] = 'Test location 1'
- WHERE [Name] = 'Test location'
+SET [Name] = 'Test location 1'
+WHERE [Name] = 'Test location'
 GO
 
 DELETE FROM [Production].[Location]
 WHERE [Name] = 'Test location 1';
 
-SELECT * FROM [Production].[LocationHost];
+SELECT *
+FROM [Production].[LocationHost];
